requirements.txt
pandas
python-dotenv
neo4j
huggingface_hub
sentence_transformers
scikit-learn
streamlit
numpy